OMAP3 PM: add config option for MPU L2 cache corruption workaround
Gives the option to replace MPU RETENTION state with INACTIVE state which
affects C3 and C5.  This is the default behavior.  If you disable the option
then the old behavior of allowing MPU to hit RET is active, which may cause
hangs and reboots for OMAP3 chips produced before July 2009.

Signed-off-by: Mike Turquette <mturquette@ti.com>