usb: musb: workaround for omap3 to ensure enumeration works with pm enabled
We have found that enumeration will sometimes fail or the omap3 will
enumerate as a full-speed device instead of a high-speed device when
power management is enabled and the device is transitioning in and
out of CORE retention. This workaround disables smart-idle/standby
for the musb controller when a usb cable is connected and keeps
the controller in no-idle/standby. When the usb cable is
disconnected smart-idle/standby for the musb controller is re-enabled.

This problem has been observed with OMAP3430 ES3.1 silicon on the
SDP with the Google K29 kernel.

Signed-off-by: Jon Hunter <jon-hunter@ti.com>