zoom2: remove lcd enable gpio
On Zoom2 there is no LCD enable GPIO line, earlier in the code
LCD enable was defined as 7+OMAP2_GPIO_LINES which is an error,
it suppose to be back light configuration. Also GPIO_15 was
configured as BACKLIGHT, which is an error, it is one of the
Audio lines, hence we have to include this clean patch.

This patch fixes above issues.
It is made clear that there is no LCD enable GPIO on zoom2 and
Backlight GPIO is 7.

Signed-off-by: Subramani Venkatesh <x0108988@ti.com>
Signed-off-by: Subramani Venkatesh <subramani.venkatesh@windriver.com>