imx046: Fix reported values from SENSOR_INFO ioctl
IMX046 uses a private ioctl to return data about the current
base size and XClock.
This patch changes the reported info so that user space can calculate
the frame rate and what is the timing of the active area.

For each base size the foloowing is now reported:

 Full size:    Top = 0, Left = 0, Width = 3440, Height = 2510
 Active size:  Top = 0, Left = 0, Width = 3280, Height = 2464

 Full size:    Top = 0, Left = 0, Width = 3456, Height = 1250
 Active size:  Top = 0, Left = 0, Width = 1280, Height = 720

 Full size:    Top = 0, Left = 0, Width = 3440, Height = 629
 Active size:  Top = 0, Left = 0, Width = 820, Height = 616

 Full size:    Top = 0, Left = 0, Width = 3440, Height = 629
 Active size:  Top = 0, Left = 0, Width = 410, Height = 308

The clock value returned has been changed to return the Pixel Clock
instead of XClk.

Signed-off-by: Rado <ribrishimov@mm-sol.com>