zoom2: enabling conservative and userspace cpu freq governors
The CPUfreq governor "conservative", much like the "ondemand" governor,
sets the CPU depending on the current usage.  It differs in behaviour
in that it gracefully increases and decreases the CPU speed rather than
jumping to max speed the moment there is any load on the CPU

The CPUfreq governor "userspace" allows the user, or any userspace
program running with UID "root", to set the CPU to a specific frequency
by making a sysfs file "scaling_setspeed" available in the CPU-device
directory.

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>