OMAP3: Disable Autoidle bit for Interrupt Controller
If the Autoidle bit of a Module is set it will shut off its
Interface clock if the module does not place any request on the bus
thereby saving power. According to INTH AUTOIDLE Issue Interrupt
controller is not able to detect interrupts due to autoidle bit  being
set and thus the OCP clock is disabled so the incoming interrupts
are not sampled.

The software workaround is to disable the Autoidle bit before WFI
situation and to enable it afterwards.

Signed-off-by: Shweta Gulati <shweta.gulati@ti.com>
Signed-off-by: Vishwanath Sripathy <vishwanath.bs@ti.com>