omap3isp: Fix constraints for resizer ratio
The test for the upscaling resizer ratio was not applying
the constraint correctly.
For example a resize from 320x240 -> 1280x960 is not valid
according to the TRM's calculation.

HORZ = (32*sph +(ow-1)*hrsz +16) / 256 + 7

where:
  ow = 1280
  sph = 0
  hrsz = 64 (minimum value allowed)

HORZ = (32*0 +(1280-1)*64 +16) / 256 + 7
     = (0    +1279*64  +16) / 256 + 7
     = 81872 / 256  + 7
     = 319 + 7
HORZ = 326

Thus this resize requires a input width of 326 pixels, but in our
example the input width is only 320 !

This patch fixes the test for the out of bounds resize ratio and
returns -EINVAL if it is out of bounds.

Signed-off-by: Dominic Curran <dcurran@ti.com>
(cherry picked from commit b67737aea48684e170de216a7798ae914488a40c)