omap3isp: Reshuffle interrupt handling
NOTE: This is based on original patch by Sakari Ailus of the same
      name. Adapted to work without prior code cleanup.

Original description:

LSC_PRE_ERR, CSIA and IRQ0STATUS_CSIB_IRQ interrupts are not handled when
the image data is ignored at the beginning of streaming. The CATCHALL
handler is not called either in that case.

Now there is always a video buffer so we can get the pointer (buf) in the
beginning of the handler.

Signed-off-by: Sakari Ailus <sakari.ailus@maxwell.research.nokia.com>
Signed-off-by: Sergio Aguirre <saaguirre@ti.com>