OMAP3630: Clock: add clksel_shift to struct clk
Introduces clksel_shift to struct clk and populates the same for dpll4_m3_ck,
dpll4_m4_ck and dpll4_m5_ck.  The purpose of this change is to make it easy
for a generic function to refresh specific CM_CLKSEL registers on a per-clock
basis by first adding then subtracting the shift value from the register.
This is needed in 3630 to work around a bug in the new j-type DPLL4.

Signed-off-by: Mike Turquette <mturquette@ti.com>