USB: MUSB: Fix the possible kernel panic while resuming
While waking up, musb can cause a kernel panic. This patch
fixes this by enabling the clock in the resume_early method.

If USB is suspended with interrupt enabled and connected to
host PC, kernel panic would occur when it wakes up. Because,
after the arch_suspend_enable_irqs() function is called in
the suspend_enter() function, USB Interrupt handler is called,
even though USB controller is still not resumed.

Problem seems to occur because we resume our clocks in the
musb_resume() function which is called after the interrupts
are enabled. Instead, we should do it in resume_early instead.

This patch was tested on Zoom2 and the problem is gone.

Signed-off-by: Kim Kyuwon <chammoru@gmail.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>