omap: serial: set correct sw flow ctrl mode
This Patch makes sure that the correct SW mode is Set in the
UART_EFR register according to the termios c_iflag.

It clears the 4 lower bits(SW flow control mode) of what
we read from the UART_EFR.

Previously, the new UART_EFR setting was getting or'd with
the original value and was not setting the correct sw flow
control mode. As A result, sometimes the SW mode was 0x1,
which meant it would read from X0FF2 To halt communication.
Since XOFF2 is not set, The terminal would stop responding
with some random Character.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>