USB: MUSB: Do not enable TX and RX DMA at the same time
This is a workaround for OMAP3 errata 1.130.

This bug causes an occassional DMA controller hang when
two DMA channels are simultaneously enabled in RX and TX
directions.

With this workaround, in the best case, we effectively use
DMA in both directions, and in the worst case, we use PIO
mode in one direction. The average throughput is better
than using PIO mode in one direction all the time.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Nilkesh Patra <Nilkesh.patra@ti.com>
Signed-off-by: Ajay Kumar Gupta <ajay.gupta@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>