OHCI: force the domain to Idle/Standby while disabling the f-clocks
If the USB F-clocks are disabled while the domain is active, we
will not be able to activate the domain again. Workaround this
by forcing the domain to Idle/Standby while disabling the f-clocks.

The root cause of this issue is not yet known, so this is a
temporary workaround to prevent a kernel panic.

We force-idle, force-standby the controller before disabling the
f-clocks and put it back to no-idle, no-standby after enabling
the f-clocks.

Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Harish <h-harish@ti.com>
(cherry picked from commit b555e81a64faceeb765984bd79b948edb5b81e48)