omap3isp: Fix func call order in isp_tmp_buf_alloc()
If the temp buffer size is bigger then the requested size
then we do not reallocate the buffer (for performance reason).

However, we still need to make sure we update the image
offset (stride) to account for new image size.
Thus the call to ispresizer_config_inlineoffset() has been
moved above the test for the buffer size.

Signed-off-by: Dominic Curran <dcurran@ti.com>