USB: MUSB: Fix DMA allignment in the RX path
On the new MUSB controller (RTL rev 1.8 on OMAP3630 and
RTL rev 2.0 on OMAP3430), DMA is not working properly if
the data in the destination buffer is not alligned on 4
byte address.

This patch fixes the DMA allignemt in the RX path by
over-riding the NET_IP_ALIGN macro to 0 bytes, to have
destination buffer aligned at 4 bytes thereby getting
alligned addess for DMA access.

Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
Signed-off-by: Anand Gadyar <gadyar@ti.com>