Call flush_dcache_page after PIO data transfer in libata-aff.c
When reading data from an ATA device using PIO, the kernel dirties the
D-cache but there is no flush_dcache_page() call in ata_pio_sector().
Since neither the VFS layer calls this function, a subsequent
update_mmu_cache() is not aware of the dirty page which may lead to
cache incoherency in user space.

Signed-off-by: Catalin Marinas <catalin.marinas@arm.com>