UART: Restore MCR register value on wake up
Fixes issue related to the UART console corruption on ZOOM3
boards on exiting Sleep Mode. The UART console gets corrupted
only when enable_off_mode is enabled in the init.rc file. The
Enable Off Mode, shuts down the CPU while in sleep mode.

On exiting the sleep mode, since the MCR register of the QUART
is not getting set properly.

The Fix is in "serial_omap_set_mctrl", where mcr register write
needs to follow sequence setting of EFR[4] to write the MCR[7..5]
bits. The Xon bit needs to be restored for the QUART proper
operation

Signed-off-by: Deepak K <deepak.k@ti.com>