OMAP3:PM: DPLL4 autoidle handling across OsWR
Currently OSWR does not
work if rom code restores DPLL4 to non auto idle mode.
ROM restore takes 20mS longer if PER/DPLL4 idle is enabled before OFF.
So it is typically not enabled. Since OSWR hangs if it is not enabled
enable it for OSWR alone. Later in the restore path
it is disabled again.

Need to REVISIT, it's more like a hack.

Signed-off-by: Girish S G <girishsg@ti.com>
Signed-off-by: Thara Gopinath <thara@ti.com>