OMAP3:DSS: enable clk at the right place
There is a possible bug, where in if restore_all_ctx is executed then
clks are disabled while the driver still expects it to be enabled.
This would cause a data abort.

To avoid, place the enabling of clks after restore in the code path.

This was found on one of the use case on custom board, where the suspend
breaks in between but the DSS has hit OFF already.

Signed-off-by: Girish S G <girishsg@ti.com>