omap34xxcam: Remove requeueing of buffer on error
This remove code to requeue a buffer when it has been dequeued
with an error.

HACK: This change has been made to stop the image rolling when
the 3630 temp buffer removal patch is applied.
This is patch is temporary and will be reverted once issues with
the temp buffer removal patch have been resolved.

Signed-off-by: Dominic Curran <dcurran@ti.com>