OMAP3: CPUidle: Update statistics for correct state
When 'enable_off_mode' is 0, the target power state for MPU
and Core is locally changed to PWRDM_POWER_RET but, the
statistics are updated for idle state originally selected
by the governor.

This patch 'invalidates' the idle states that lead either of
MPU or Core to PWRDM_POWER_OFF state when 'enable_off_mode'
is '0'. The states are valid once 'enable_off_mode' is set
to '1'.

This patch has been ported from L-O, original patch can be found here:
http://patchwork.kernel.org/patch/62533/

Signed-off-by: Sanjeev Premi <premi@ti.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>