OMAP: Powerdomains: Add support for INACTIVE state on pwrdm level
Currently only ON, RET and OFF are supported, and ON is arguably
broken as it allows the powerdomain to enter INACTIVE state unless
idle is prevented. Now, pwrdm code prevents idle if ON is selected,
and also adds support for INACTIVE. This simplifies the control
needed inside sleep code.

This patch has been ported from L-O, original patch can be found here:
http://patchwork.kernel.org/patch/59601/

Signed-off-by: Tero Kristo <tero.kristo@nokia.com>
Signed-off-by: Vishwanath BS <vishwanath.bs@ti.com>