ARM: OMAP: Add API function to enable prefetch on DMA channel
- This feature improves latency by fetching data before receiving
   the DMA request (note that the data shall be already available).
 - Previous API did not allow to enable it.
 - Default behavior is NOT changed if the new optional API is not called.

Previous API did not manage prefetch (disable by default). Prefetch has been
introduced in DMA4 (not applicable to OMAP1).
Documentation specifies:
"The performance of a hardware-synchronized transfer can be improved by using
the prefetch mode, enabled through the channel DMA register bit
SDMA.DMA4_CCRi[23]. Data is prefetched on the read port side in advance of the
DMA request received and buffered in the FIFO queue. Up to a full transfer
block can be prefetched, although this can be limited by the specified maximum
channel FIFO queue depth."

Prefetch only applies to destination synchronized transfer.
However, prefetch shall be disabled if the source data may not be
available when the DMA channel is enabled.
Behavior is undefined when prefetch is enabled and a transfer is synchronized
to the source.

This patch has been tested on L25.12 with a custom SPI driver used to communicate
with a 3G modem (prototyping). The latency reduction has not been evaluated yet.

Signed-off-by: Patrick Combes <p-combes@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>