sil9022: Fix for sil9022_probe() when sil9022 is not present on the board
sil9022_probe was not failing cleanly when sil9022 is not present on board.
The probe should fail normally and not result in crash. The crash was a result
of releasing the wrong variable in place of sil9022_client. this fix takes
care of this crash.

Signed-off-by: Srinivas Pulukuru <srinivas.pulukuru@ti.com>
Acked-by: Mikkel Christensen <mlc@ti.com>
Acked-by: Leed Aguilar <leed.aguilar@ti.com>
(cherry picked from commit 00700bf3f5609e533e0663abad9947e4e24dc3bc)