TWL4030: Battery: Notify user of USB cable plugged.
USB plug-in does not generate a interupt on the
battery Driver. The USB interrupt is handeled
by the USB driver which calls the usb_en function
in this battery driver.

Since in usb_en we are not calling power_supply_changed
to notify user space of the plug in event, We wait
until the next update cycle to detect the change in
power_suply and send the event to user space.

Signed-off-by: Axel Haslam <axelhaslam@ti.com>