TWL4030: Battery: Add sysfs entry for USB battery
Android battery service requires a sysfs folder on
/sys/class/power_supply that has a battery with the
"type" property POWER_SUPPLY_TYPE_USB.
The battery service uses the "online" property of
this battery to determine if the USB cable is
plugged in or not.

In turn, The USB activity Service is notified
by the battery service if the cable is plugged.

Without this sysfs entry, USB activity would
not get properly notified of the status of the
USB cable.

Change-Id: I1c1e704fc589735db68c49a1649a988da7982d7e
Signed-off-by: Axel Haslam <axelhaslam@ti.com>
Acked-by: Madhusudhan Chikkature <madhu.cr@ti.com>