zoom: camera: constrain mpu to C1 while camera power on
Add a constraint to keep mpu in C1 state while camera is on.

This fixes a problem seen when mpu goes to active state.
The active state latency _may_ be a problem for camera.

This patch should be used in tandem with the following pm patch:
OMAP3: PM: cpuidle deny INA state for MPU and CORE in C1

Fundamental issue needs further attention !

Signed-off-by: Dominic Curran <dcurran@ti.com>
Acked-by: Girish S G <girishsg@ti.com>