hack: mmc: Call flush_dcache_page in the vfs
When reading data from an MMC device , the kernel dirties the
D-cache. This patch adds l2 cache flush along with existing L1 cache.
This issue is mainly because of cache-policy used on SMP systems.

This should be fixed at MMC stack level and but so far I haven't
located the right place to put this fix.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Tested-by: Madhusudhan Chikkature <madhu.cr@ti.com>