DSPBRIDGE: Fix blank screen issue
This patch fixes the blank screen seen while pressing any key to
wakeup from suspend. GPTimer8 is being used for MMUFault notification;
currently, dm timer framwork automatically enables the clock when
requesting it, but this is not needed in this case as it will be used
for a short period of time, so disabling it after request.

Signed-off-by: Ernesto Ramos <ernesto@ti.com>
Signed-off-by: Shivananda Hebbar <x0hebbar@ti.com>