OMAP3:PM: DPLL3 RTL bug Errata i570 WA
When internal SWAKEUP event occurs while CORE DPLL is going to bypass,
PRCM may report that DPLL3 is locked whereas it is not.
This is due to the timing path of an internal signal which is not fitting
to one L3 clock cycle.

In that case, PRCM DPLL control FSM deasserts SDRC_IDLEREQ signal before
DPLL3 Lock state is set.

Consequence is that SDRC is released from IDLE with bypass clock
(which is too low) instead of the locked frequency. DLL may/maynot lock based
on Process Voltage Temperature conditions.

WA refer to Errata i570/i571
For 3430:
Before RET/OFF:
	1) Reduce DPLL3 M2 Frequency to get L3 running at OPP2 Frequency
	2) Increase CORE Voltage to 1.2V
	3) Enable DPLL3 Automatic mode. This will ensure proper transition to
	   RETENTION or OFF mode.
After RET/OFF:
	1) Disable DPLL3 Automatic mode.
	2) Restore previous DPLL3 M2 Frequency and CORE Voltage values.
For 3630:
Before RET/OFF:
	1) Reduce DPLL3 M2 Frequency to get L3 running at OPP50 Frequency
	2) Keep volt at 1.1375V
	3) Enable DPLL3 Automatic mode.
After:
	1) Disable DPLL3 Automatic mode
	2) Restore previous DPLL3 M2 Frequency and CORE Voltage values.

Validate 3630/3430. Scoped VDD2 signals are behaving as expected.

Signed-off-by: Girish S G <girishsg@ti.com>