DSPBRIDGE: Improvement to trace buffer dump on DSP crash
This fix it will print the trace buffer content in the proper order.
Prior to this, the contents were dumped without regards of the
wrap-around point of circular buffer. As a result newer traces
would print before older ones and the last printed before a
crash would not always be apparent.

Signed-off-by: Cris Jansson <cjansson@ti.com> [change to open source coding style]
Signed-off-by: Ernesto Ramos <ernesto@ti.com>
Signed-off-by: Shivananda Hebbar <x0hebbar@ti.com>