OMAP3630: revert the FBB/ABB implementation for 1GHz
This patch reverts the following patch series:

- OMAP3: PM: SR adjustment of test nvalue for OPP4
- OMAP3630: PM: implement Foward Body-Bias for OPP1G
- OMAP3630: PM: update PRM registers for 3630
- Revert "OMAP3:PM:DSP: Don't allow max ARM freq when DSP load is high"
- Revert "OMAP3:PM: don't do multiple lock on VDD1 when DSP request for"

The above patch series was reverted since it requires more testing
in order to be included in the L25.inc2.5 patch 3 release. There is
nothing wrong with it, but we must assure a complete validation across
the whole system.

To add these feature/fixes again, we just need to revert this commit ID

Notice that this patch is bringing back the HACK for not allowing the
ARM going to 1GHz when the DSP load is high.

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>